[refactor] 인증 로직 통합
