--- BE/promate/src/main/java/org/example/promate/domain/project/respository/ProjectRepository.java
@@ -1,12 +1,0 @@
-package org.example.promate.domain.project.respository;
-
-import org.example.promate.domain.project.entity.Project;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface ProjectRepository extends JpaRepository<Project, Long> {
-    Optional<Project> findById(Long id);
-}
