+++ BE/promate/src/main/java/org/example/promate/domain/workspace/service/query/ScheduleQueryServiceImpl.java
@@ -3,7 +3,7 @@ package org.example.promate.domain.workspace.service.query;
 import lombok.RequiredArgsConstructor;
 import org.example.promate.domain.project.code.MemberErrorCode;
 import org.example.promate.domain.project.exception.MemberException;
-import org.example.promate.domain.project.respository.MemberRepository;
+import org.example.promate.domain.project.repository.MemberRepository;
 import org.example.promate.domain.workspace.code.ScheduleErrorCode;
 import org.example.promate.domain.workspace.converter.ScheduleConverter;
 import org.example.promate.domain.workspace.dto.res.ScheduleResDto;
@@ -26,7 +26,7 @@ public class ScheduleQueryServiceImpl implements ScheduleQueryService{
     @Override
     public ScheduleResDto.MonthlyScheduleListInfoDto getMonthlySchedules(Long userId, Long projectId, Integer year, Integer month) {
         //사용자가 해당 프로젝트의 팀원인지 확인
-        if(!memberRepository.existsByUser_IdAndProject_Id(userId, projectId)){
+        if(!memberRepository.existsByUserIdAndProjectId(userId, projectId)){
             throw new MemberException(MemberErrorCode.SCHEDULE_FORBIDDEN_NOT_PROJECT_MEMBER);
         }
 
@@ -45,7 +45,7 @@ public class ScheduleQueryServiceImpl implements ScheduleQueryService{
     @Override
     public ScheduleResDto.ProjectScheduleDetailInfoDto getScheduleDetails(Long userId, Long projectId, Long scheduleId) {
         //사용자가 해당 프로젝트의 팀원인지 확인
-        if(!memberRepository.existsByUser_IdAndProject_Id(userId, projectId)){
+        if(!memberRepository.existsByUserIdAndProjectId(userId, projectId)){
             throw new MemberException(MemberErrorCode.SCHEDULE_FORBIDDEN_NOT_PROJECT_MEMBER);
         }
